@@ -23,23 +23,23 @@
     "gantt"
   ],
   "dependencies": {
-    "config": "^3.3.1",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
-    "ejs": "^3.0.1",
-    "express": "~4.16.0",
-    "http-errors": "~1.6.2",
-    "moment": "^2.26.0",
+    "config": "^4.3.0",
+    "cookie-parser": "^1.4.7",
+    "debug": "^4.4.3",
+    "ejs": "^4.0.1",
+    "express": "^5.2.1",
+    "http-errors": "^2.0.1",
+    "moment": "^2.30.1",
     "moment-business-days": "^1.2.0",
-    "morgan": "~1.9.0"
+    "morgan": "^1.10.1"
   },
   "engines": {
-    "node": ">=12.0"
+    "node": ">=24.13.0"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
-    "lodash": "^4.17.21",
-    "mocha": "^7.2.0",
-    "sinon": "^9.0.2"
+    "chai": "^6.2.2",
+    "lodash": "^4.17.23",
+    "mocha": "^11.7.5",
+    "sinon": "^21.0.1"
   }
 }
