@@ -1,13 +1,12 @@
 import React, { useEffect } from "react";
-import { Route, Switch } from "react-router-dom";
-import "carbon-components/css/carbon-components.min.css";
-import "./App.css";
-import { Loading } from "carbon-components-react";
-import { HqPageAll, HqPageTag, HqPageSwagger } from "./pages";
-import { HqFooter, HqHeader, HqTagList } from "./components";
-import { useInterval } from "./utils/useInterval";
+import { Route, Routes } from "react-router";
+import "./App.scss";
+import { Loading } from "@carbon/react";
+import { HqPageAll, HqPageTag, HqPageSwagger } from "/src/pages";
+import { HqFooter, HqHeader, HqTagList } from "/src/components";
+import { useInterval } from "/src/utils/useInterval";
 import Headroom from "react-headroom";
-import apiData from "./test/mock-api-data";
+import apiData from "/src/test/mock-api-data";
 
 function App() {
   
@@ -18,7 +17,7 @@ function App() {
     React.useState("1970-01-01");
 
   const makeRequest = () => {
-    if (process.env.REACT_APP_DEMO !== "true") {
+    if (import.meta.env.VITE_DEMO !== "true") {
       fetch("/todos/filelastupdated")
         .then((res) => res.json())
         .then((currentSourceTimestamp) => {
@@ -38,7 +37,7 @@ function App() {
   };
 
   useEffect(() => {
-    if (process.env.REACT_APP_DEMO === "true") {
+    if (import.meta.env.VITE_DEMO === "true") {
       setData(apiData);
       setIsLoading(false);
     }
@@ -53,25 +52,19 @@ function App() {
   }
 
   return (
-    <div>
+    <div>      
       <Headroom>
         <HqHeader userName={data.whoami} menuLinks={data.menuLinks} />
         <HqTagList tags={data.tags} />
       </Headroom>
-      <Switch>
-        <Route path="/" exact>
-          <HqPageAll data={data} />
-        </Route>
-        <Route path="/tags/:tag">
-          <HqPageTag data={data} />
-        </Route>
-        <Route path="/doc">
-          <HqPageSwagger />
-        </Route>
-      </Switch>
+      <Routes>
+        <Route path="/" exact element={<HqPageAll data={data} />} />
+        <Route path="/tags/:tag" element={<HqPageTag data={data} />} />
+        <Route path="/doc" element={<HqPageSwagger />} />
+      </Routes>
       <HqFooter
         pageUpdated={
-          process.env.REACT_APP_DEMO !== "true" ? data.pageupdated : new Date().toISOString()
+          import.meta.env.VITE_DEMO !== "true" ? data.pageupdated : new Date().toISOString()
         }
         fileUpdated={data.fileupdated}
       />
