@@ -1,3 +1,3 @@
-const apiDocData = process.env.REACT_APP_DEMO !== "true" ? {} : {"title":"hqTodo server app","routes":[{"method":"GET","path":"/todos/filelastupdated","description":"Timestamp in milliseconds when the TODO file was last modified.\n  For client polling to check if the file data needs to be reloaded."},{"method":"GET","path":"/todos/raw","description":"Entries from TODO file (or cache), filtered to only archived, open,\n  and taginfo."},{"method":"GET","path":"/todos/unordered","description":"Open items, with all metadata except gantt ordering info."},{"method":"GET","path":"/todos","description":"Open items, with interrupt and ordering for mermaid."},{"method":"GET","path":"/todos/archived","description":"Archived items, with metadata extracted"},{"method":"GET","path":"/todos/tags","description":"(1) title and link for tags in Taginfo section of TODO file.\n  (2) list of all tags, with classs and counts."},{"method":"GET","path":"/todos/tags/:tagname","description":"The active, done, and archived entries and chart data for the specifiec tag."},{"method":"GET","path":"/todos/searchdata","description":"All the things, with tag array and which state it's in."},{"method":"GET","path":"/todos/supportdata","description":"All support entries, with time in minutes."},{"method":"GET","path":"/todos/supportondate/:ondate","description":"Support item list and total time string for a given date."},{"method":"GET","path":"/todos/archivedbytag","description":"Archived items, grouped by tag with group est totals."},{"method":"GET","path":"/todos/archivedbyweek","description":"Archived items, grouped by week (ending Sunday)"},{"method":"GET","path":"/ejs/tags/:tagname","description":"All active, done, archived, and chartdata for given tag.\n  (server-side view rendering)"},{"method":"GET","path":"/ejs","description":"Classic server-rendered view of all the things"},{"method":"GET","path":"/webc","description":"Half-started Web component rendering of all the stuff.\n  Project stopped in favor of React."},{"method":"GET","path":"/api","description":"All the things, and I mean _all_, for consumption by React client.\n  TODO: Determine where you want what for optimum client-server efficiency,\n  pushing some of the swizzling to the client, and prune this data accordingly."},{"method":"GET","path":"/api/tags/:tagname","description":"All active, done, archived, and chartdata for given tag.\n  (server-side swizzling for client rendering)\n  (deprecated - swizzling done on client now)"},{"method":"GET","path":"/api/doc","description":"Documentation data for the Express server portion of hqTodo."},{"method":"GET","path":"*","description":"All other routes handled by the client application."}]};
+const apiDocData = import.meta.env.VITE_DEMO !== "true" ? {} : {"title":"hqTodo server app","routes":[{"method":"GET","path":"/todos/filelastupdated","description":"Timestamp in milliseconds when the TODO file was last modified.\n  For client polling to check if the file data needs to be reloaded."},{"method":"GET","path":"/todos/raw","description":"Entries from TODO file (or cache), filtered to only archived, open,\n  and taginfo."},{"method":"GET","path":"/todos/unordered","description":"Open items, with all metadata except gantt ordering info."},{"method":"GET","path":"/todos","description":"Open items, with interrupt and ordering for mermaid."},{"method":"GET","path":"/todos/archived","description":"Archived items, with metadata extracted"},{"method":"GET","path":"/todos/tags","description":"(1) title and link for tags in Taginfo section of TODO file.\n  (2) list of all tags, with classs and counts."},{"method":"GET","path":"/todos/tags/:tagname","description":"The active, done, and archived entries and chart data for the specifiec tag."},{"method":"GET","path":"/todos/searchdata","description":"All the things, with tag array and which state it's in."},{"method":"GET","path":"/todos/supportdata","description":"All support entries, with time in minutes."},{"method":"GET","path":"/todos/supportondate/:ondate","description":"Support item list and total time string for a given date."},{"method":"GET","path":"/todos/archivedbytag","description":"Archived items, grouped by tag with group est totals."},{"method":"GET","path":"/todos/archivedbyweek","description":"Archived items, grouped by week (ending Sunday)"},{"method":"GET","path":"/ejs/tags/:tagname","description":"All active, done, archived, and chartdata for given tag.\n  (server-side view rendering)"},{"method":"GET","path":"/ejs","description":"Classic server-rendered view of all the things"},{"method":"GET","path":"/webc","description":"Half-started Web component rendering of all the stuff.\n  Project stopped in favor of React."},{"method":"GET","path":"/api","description":"All the things, and I mean _all_, for consumption by React client.\n  TODO: Determine where you want what for optimum client-server efficiency,\n  pushing some of the swizzling to the client, and prune this data accordingly."},{"method":"GET","path":"/api/tags/:tagname","description":"All active, done, archived, and chartdata for given tag.\n  (server-side swizzling for client rendering)\n  (deprecated - swizzling done on client now)"},{"method":"GET","path":"/api/doc","description":"Documentation data for the Express server portion of hqTodo."},{"method":"GET","path":"*","description":"All other routes handled by the client application."}]};
 
 export default apiDocData;
